Move podschedulers below components/algorithm
