+++ contrib/mesos/pkg/scheduler/errors/errors.go
@@ -25,4 +25,5 @@ var (
 	NoSuchPodErr               = errors.New("No such pod exists")
 	NoSuchTaskErr              = errors.New("No such task exists")
 	ReconciliationCancelledErr = fmt.Errorf("explicit task reconciliation cancelled")
+	NoSuitableOffersErr        = errors.New("No suitable offers for pod/task")
 )
